Implement agent function calling
With this commit, the agent is now capable of actually calling functions
through the call_function() method, returning the function call reponse
to generate_content() or raising a fatal error if a function call
returns no/invalid response. If an unrecognized function name is passed
to call_function, it returns an error reponse to the agent but does not
crash the program like an invalid response to a valid function name
will.