Implement feedback loop for agent iteration
This commit refactors generate_content() to include a loop with a
maximum iteration of 20 passes that can call the LLM over and over,
while also maintaining a list of messages in the "conversation" to keep
track of progress. Each iteration checks the .candidates property of the
response, adding the candidate.content to the messages list.
Additionally, each function call now appends the types.Content reponse
from the function to messages as well. After each loop, if a function
was called, it iterates again (unless the maximum iterations was
reached). Otherwise, the LLM's final response is printed to output --
meaning the agent is done with the task (or failed).